tool: Adapt to Deepin 20.x environment

+++ setup.sh
@@ -10,6 +10,9 @@ cd $LOCALDIR
 USE_MIRROR_FOR_PIP=true
 # Python pip mirror link
 PIP_MIRROR=https://pypi.tuna.tsinghua.edu.cn/simple/
+# OS_ID delect
+OS_ID=$(sudo cat /etc/os-release | awk -F "=" '{if ($1 == "ID") print $2}')
+[ -z "$OS_ID" ] && OS_ID=$(sudo cat /etc/os-release | awk -F "=" '{if ($1 == "ID_LIKE") print $2}')
 
 dump_welcome(){
     echo -e "\033[34m $(cat banner/banner) \033[0m"
@@ -23,7 +26,6 @@ dependency_install(){
     if [[ "$OSTYPE" == "linux-gnu" ]]; then
         echo -e "\033[33m [DEBUG] Linux Detected \033[0m"
         echo -e "\033[33m [INFO] Installing Packages via apt... \033[0m"
-        sudo apt update && sudo apt upgrade -y
         sudo apt install git p7zip openjdk-8-jdk curl cpio wget unace unrar zip unzip p7zip-full p7zip-rar sharutils uudeview mpack arj cabextract file-roller aptitude device-tree-compiler liblzma-dev liblz4-tool gawk aria2 selinux-utils busybox bc -y
         sudo apt update --fix-missing
         
@@ -36,24 +38,29 @@ dependency_install(){
 
 python_install(){
         echo -e "\033[33m [INFO] Python and Pip install... \033[0m"
-        sudo apt-get --purge remove -y python3-pip
-        sudo apt install python aptitude -y
-        sudo aptitude install python-dev -y
-        sudo add-apt-repository universe
-        sudo python get-pip.py
+        sudo apt install python python-pip -y
+        sudo python -m pip install --upgrade pip
+        if [[ $OS_ID == "ubuntu" ]]; then
+            sudo apt-get --purge remove -y python3-pip
+            sudo apt install python aptitude -y
+            sudo aptitude install python-dev -y
+            sudo add-apt-repository universe
+            sudo python get-pip.py
+        fi
         sudo apt install python3 python3-pip -y
+        sudo python3 -m pip install --upgrade pip
 }
 
 pip_module_install(){
     echo -e "\033[33m [INFO] Python2 and Python3 module install... \033[0m"
     if [[ "$USE_MIRROR_FOR_PIP" == "true" ]] ; then
         echo -e "\033[33m [INFO] Installing python packages from mirror... \033[0m"
-        sudo pip install backports.lzma pycryptodome pycrypto -i $PIP_MIRROR
-        sudo pip3 install backports.lzma pycryptodome pycrypto -i $PIP_MIRROR
+         pip install backports.lzma pycryptodome pycrypto -i $PIP_MIRROR
+         pip3 install backports.lzma pycryptodome pycrypto -i $PIP_MIRROR
     elif [[ "$USE_MIRROR_FOR_PIP" == "false" ]] ; then
         echo -e "\033[33m [INFO] Installing python packages from python-pip offical... \033[0m"
-        sudo pip install backports.lzma pycryptodome pycrypto
-        sudo pip3 install backports.lzma pycryptodome pycrypto
+        pip install backports.lzma pycryptodome pycrypto
+        pip3 install backports.lzma pycryptodome pycrypto
     fi
     
     if [[ "$USE_MIRROR_FOR_PIP" == "true" ]] ; then
@@ -66,9 +73,9 @@ pip_module_install(){
     elif [[ "$USE_MIRROR_FOR_PIP" == "false" ]] ; then
         for requirements_list in $(find $LOCALDIR -type f | grep "requirements.txt");do
             echo -e "\033[33m [INFO] from requirements.txt Installing python packages \033[0m"
-            sudo pip install -r $requirements_list
+            pip install -r $requirements_list
             echo -e "\033[33m [INFO] from requirements.txt Installing python3 packages \033[0m"
-            sudo pip3 install -r $requirements_list
+            pip3 install -r $requirements_list
         done
     fi
 }
